Database settings in django has been moved to yml
As we are working on separate databases on different machines some settings
may not work both machines. `settings_databases.yml` is pushed so that it
can be used as a mould to make your custom database config. Please use -

```
git update-index --assume-unchanged BookKeep/settings_databases.yml
```

after pulling so that the changes you make do not affect work tree